# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_handlers.EmailLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  internal_handlers.SignupRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  internal_handlers.SteamLoginRequest:
    properties:
      steam_id:
        type: string
    type: object
  internal_handlers.UpdateEmailRequest:
    properties:
      email:
        type: string
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: The Game Journal (Letterboxd for games) API
  title: Checkpoint MVP API
  version: "1.0"
paths:
  /auth/email:
    put:
      consumes:
      - application/json
      description: Update the user's email address
      parameters:
      - description: Update Email Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.UpdateEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update Email
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user using email and password
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.EmailLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Traditional Login
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Register a new user with email and password
      parameters:
      - description: Signup Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.SignupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Traditional Signup
      tags:
      - auth
  /auth/steam:
    post:
      consumes:
      - application/json
      description: Authenticates or creates a user using their Steam ID and syncs
        library
      parameters:
      - description: Steam Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.SteamLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login/Signup with Steam
      tags:
      - auth
  /auth/steam/link:
    post:
      consumes:
      - application/json
      description: Link a steam account to an existing user and sync library
      parameters:
      - description: Steam Link Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.SteamLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Link Steam Account
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"